Move naflex data config to base TrainingTask so it can be used by audio tasks now
